Initial (very crude) use of the component library
This is a crude version of creating components (as defined by Stuart
Sierra's Component library, https://github.com/stuartsierra/component)
from KS topologies. At the moment, it's possible to start the
components from the CLI and have them work together. The goal is to
easily run the topologies (in separate containers) from the REPL.